+++ src/__tests__/integration.js
@@ -9,23 +9,20 @@ const config = {
     "block-opening-brace-newline-after": "always",
     "declaration-block-properties-order": [
       {
-        emptyLineBefore: "always",
         properties: [
           "content",
         ],
       },
       {
-        emptyLineBefore: "always",
         properties: [
-          "position",
-          "top",
-          "right",
-          "bottom",
-          "left",
-          "z-index",
+          "display",
+          "height",
         ],
       },
     ],
+    "declaration-property-unit-blacklist": {
+      "width": [ "px", "em" ],
+    },
     "color-no-invalid-hex": [ true, {
       severity: "warning",
       message: "You made a mistake",
@@ -37,11 +34,15 @@ const config = {
 }
 
 const css = (
-`a {
+`
+a { background: pink; }
+
+b {
+  height: 1rem;
+  display: block;
+  width: 10px;
   color: #zzz;
 }
-
-b { background: pink; }
 
 /* stylelint-disable color-no-invalid-hex */
 .foo {
@@ -51,11 +52,12 @@ b { background: pink; }
 
 .bar:before {
   color: #mmm;
+  height: calc(1rem - 100%)
 }
 `)
 
 test("integration test expecting warnings", t => {
-  t.plan(9)
+  t.plan(15)
 
   postcss()
     .use(stylelint(config))
@@ -65,15 +67,31 @@ test("integration test expecting warnings", t => {
 
   function checkResult(result) {
     const { messages } = result
-    t.equal(messages.length, 3)
+    t.equal(messages.length, 6)
     t.ok(messages.every(m => m.type === "warning"))
     t.ok(messages.every(m => m.plugin === "stylelint"))
+
+    // block-opening-brace-newline-after - string primary option
     t.equal(messages[0].text, "Expected newline after \"{\" (block-opening-brace-newline-after)")
     t.equal(messages[0].severity, "error")
-    t.equal(messages[1].text, "You made a mistake")
-    t.equal(messages[1].severity, "warning")
-    t.equal(messages[2].text, "You made a mistake")
-    t.equal(messages[2].severity, "warning")
+
+    // declaration-block-properties-order - nested array primary option
+    t.equal(messages[1].text, "Expected \"display\" to come before \"height\" (declaration-block-properties-order)")
+    t.equal(messages[1].severity, "error")
+
+    // declaration-property-unit-blacklist - object primary option
+    t.equal(messages[2].text, "Unexpected unit \"px\" for property \"width\" (declaration-property-unit-blacklist)")
+    t.equal(messages[2].severity, "error")
+
+    // color-no-invalid-hex - true primary option
+    t.equal(messages[3].text, "You made a mistake")
+    t.equal(messages[3].severity, "warning")
+    t.equal(messages[4].text, "You made a mistake")
+    t.equal(messages[4].severity, "warning")
+
+    // function-blacklist - array primary option
+    t.equal(messages[5].text, "Unexpected function \"calc\" (function-blacklist)")
+    t.equal(messages[5].severity, "error")
   }
 })