+++ src/rules/at-rule-name-newline-after/__tests__/index.js
@@ -51,6 +51,8 @@ testRule(rule, {
   }, {
     code: "@keyframes\nidentifier { }",
   }, {
+    code: "@-webkit-keyframes\nidentifier { }",
+  }, {
     code: "@viewport { }",
   }, {
     code: "@viewport{ }",
@@ -131,6 +133,11 @@ testRule(rule, {
     message: messages.expectedAfter("@unknown"),
     line: 1,
     column: 8,
+  }, {
+    code: "@-webkit-keyframes identifier { }",
+    message: messages.expectedAfter("@-webkit-keyframes"),
+    line: 1,
+    column: 18,
   } ],
 })
 
@@ -229,6 +236,8 @@ testRule(rule, {
     code: "@supports(animation-name: test) { }",
   }, {
     code: "@keyframes identifier { }",
+  }, {
+    code: "@-webkit-keyframes identifier { }",
   }, {
     code: "@viewport { }",
   }, {